pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adomator"
version = "0.1.0"
description = "Declarative Azure DevOps repository management — Terraform for your Azure DevOps repos"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "azure-devops>=7.1.0b4",
    "click>=8.1",
    "pydantic>=2.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock>=3.10",
]

[project.scripts]
adomator = "adomator.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["adomator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
